% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{add_and_write_projection}
\alias{add_and_write_projection}
\alias{add_and_write_pca}
\alias{add_and_write_sma}
\alias{add_and_write_pls}
\alias{add_and_write_lda}
\title{Run pca/pls/lda/sma and write results}
\usage{
add_and_write_projection(
  object,
  method,
  implementation = NULL,
  result_dir,
  geom = default_feature_plots(object),
  x = default_x(object, geom),
  color_var = default_color_var(object),
  color_values = default_color_values(object, color_var),
  shape_var = default_shape_var(object),
  group_var = default_group_var(object),
  txt_var = default_txt_var(object),
  line = default_line(object),
  na.impute = FALSE
)

add_and_write_pca(object, ...)

add_and_write_sma(object, ...)

add_and_write_pls(object, ...)

add_and_write_lda(object, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{method}{'pca', 'pls', 'lda', or 'sma'}

\item{implementation}{string}

\item{result_dir}{string: result dir}

\item{geom}{subset of \code{\link[autonomics.plot]{FEATURE_PLOTS}}}

\item{x}{string: svar mapped to x in feature plots}

\item{color_var}{string: svar mapped to color in feature plots}

\item{color_values}{named string vector: names = color_var levels, values = colors}

\item{shape_var}{string: svar mapped to shape in feature plots}

\item{group_var}{string: svar mapped to group in feature plots}

\item{txt_var}{string: svar mapped to txt   in feature plots}

\item{line}{TRUE or FALSE:  connect points in feature plots with a line?}

\item{na.impute}{TRUE or FALSE: impute missing values?}

\item{...}{for backward compatibility}
}
\value{
SummarizedExperiment with pca results
}
\description{
Run pca/pls/lda/sma and write results
}
\examples{
if (require(autonomics.data)){

   # STEM CELL COMPARISON
   require(magrittr)
   result_dir <- tempdir() \%T>\% message()
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% add_and_write_pca(result_dir = result_dir)

   # GLUTAMINASE
   object <- autonomics.data::glutaminase
   object \%>\% add_and_write_pca(result_dir = result_dir)
}

}
