% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_plots.R
\name{combine_plots}
\alias{combine_plots}
\alias{multiplot2}
\title{Combine multiple ggplots}
\usage{
combine_plots(..., plotlist = NULL, cols = 1, layout = NULL,
  file = NULL, return_object = FALSE, width = 7, height = 7)

multiplot2(...)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{ggplot objects as a list}

\item{cols}{number of columns}

\item{layout}{layout specification}

\item{file}{file to which to print to}

\item{return_object}{whether to return a grob}

\item{width}{in inches}

\item{height}{in inches}
}
\description{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
- cols:   Number of columns in layout
- layout: A matrix specifying the layout. If present, 'cols' is ignored.
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\examples{
   df <- data.frame(x = 1:100, y = 1:100)
   p <- ggplot2::ggplot(df) + ggplot2::geom_point(ggplot2::aes(x=x,y=y))
   file <- paste0(tempfile(), '.pdf')
   combine_plots(p, p)
   combine_plots(p, p, file = file)
}
\author{
Johannes Graumann
}
