% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{compute_feature_plot_dims}
\alias{compute_feature_plot_dims}
\title{Compute appropriate feature plot dims}
\usage{
compute_feature_plot_dims(p)
}
\arguments{
\item{p}{ggplot object}
}
\description{
Compute appropriate feature plot dims
}
\examples{
require(magrittr)
require(SummarizedExperiment)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios \%>\% magrittr::extract(1:9, )
   object \%>\% plot_features('point')   \%>\% compute_feature_plot_dims()
   object \%>\% plot_features('violin')  \%>\% compute_feature_plot_dims()
   object \%>\% plot_features('boxplot') \%>\% compute_feature_plot_dims()
}
}
