% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_color_values}
\alias{default_color_values}
\title{Default color values}
\usage{
default_color_values(
  object,
  color_var = default_color_var(object),
  show = FALSE
)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{color_var}{string: svar mapped to color}

\item{show}{logical}
}
\value{
default color values vector
}
\description{
Default color values
}
\examples{
if (require(autonomics.data)){

   # STEM CELL COMPARISON
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% default_color_values(show = TRUE)

   # GLUTAMINASE
   object <- autonomics.data::glutaminase
   object \%>\% default_color_values(show = TRUE)

}
}
