% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_color_var}
\alias{default_color_var}
\title{default color_var}
\usage{
default_color_var(object)
}
\arguments{
\item{object}{eset}
}
\value{
default value of color_var
}
\description{
default color_var
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   autonomics.data::stemcomp.proteinratios \%>\% autonomics.plot::default_color_var()
   autonomics.data::stemcomp.soma          \%>\% autonomics.plot::default_color_var()
}

# STEM CELL DIFFERENTIATION
if (require(billing.differentiation.data)){
  billing.differentiation.data::rna.voomcounts \%>\%
     default_color_var()
}
# ATKIN 2014
if (require(atkin.2014)){
  atkin.2014::soma \%>\% autonomics.plot::default_color_var()
}

}
