% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_fvars}
\alias{default_fvars}
\title{default fvars}
\usage{
default_fvars(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
string vector
}
\description{
default fvars
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% default_fvars()
}
}
