% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_line}
\alias{default_line}
\title{Default line value}
\usage{
default_line(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
default line value
}
\description{
Default line value
}
\examples{
if (require(autonomics.data)){
   require(magrittr)

   # STEM CELL COMPARISON
   autonomics.data::stemcomp.proteinratios \%>\% autonomics.plot::default_line()

   # GLUTAMINASE
   autonomics.data::glutaminase            \%>\% autonomics.plot::default_line()
}
}
