% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_txt_var}
\alias{default_txt_var}
\title{Default txt var}
\usage{
default_txt_var(object)
}
\arguments{
\item{object}{eset}
}
\value{
default value of txt_var
}
\description{
Default txt var
}
\examples{
require(magrittr)
}
