% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_x}
\alias{default_x}
\title{Default x}
\usage{
default_x(object, feature_plots = default_feature_plots(object))
}
\arguments{
\item{object}{eset}

\item{feature_plots}{feature plots}
}
\value{
default value of x
}
\description{
Default x
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   autonomics.data::stemcomp.proteinratios \%>\% autonomics.plot::default_x()
}

# STEM CELL DIFFERENTIATION
if (require(billing.differentiation.data)){
  billing.differentiation.data::rna.voomcounts \%>\% default_x()
}
}
