% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{extract_top_and_bottom}
\alias{extract_top_and_bottom}
\title{Extract top and bottom of sumexp}
\usage{
extract_top_and_bottom(object, n = 16)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{n}{number of top features to be selected}
}
\value{
sumex with top n/2 and bottom n/2 rows
}
\description{
Extract top and bottom of sumexp
}
