% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FEATURE_PLOTS}
\alias{FEATURE_PLOTS}
\title{FEATURE_PLOTS}
\format{
An object of class \code{character} of length 5.
}
\usage{
FEATURE_PLOTS
}
\description{
Which feature plots are supported?
}
\keyword{datasets}
