% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{gg_aeslevels}
\alias{gg_aeslevels}
\title{Get levels of variable mapped to aesthetic}
\usage{
gg_aeslevels(b, aesthetic)
}
\arguments{
\item{b}{built ggplot (result of ggplot_build)}

\item{aesthetic}{ggplot aesthetic (string)}
}
\value{
character vector
}
\description{
Get levels of variable mapped to aesthetic
}
\examples{
require(magrittr)
require(Biobase)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios \%>\% magrittr::extract(1:9, )
   b <- object \%>\% autonomics.plot::plot_features('point') \%>\%
                   ggplot2::ggplot_build()
   b \%>\% autonomics.plot::gg_aeslevels(aesthetic = 'color')
   b \%>\% autonomics.plot::gg_aeslevels(aesthetic = 'shape')
   b \%>\% autonomics.plot::gg_aeslevels(aesthetic = 'x')
}
}
