% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{gg_nrow}
\alias{gg_nrow}
\alias{gg_ncol}
\title{Get no of facet rows/columns}
\usage{
gg_nrow(b)

gg_ncol(b)
}
\arguments{
\item{b}{built ggplot (output of ggplot_build)}
}
\value{
integer
}
\description{
Get no of facet rows/columns
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016 \%>\% magrittr::extract(1:9, )
   b <- object \%>\% autonomics.plot::plot_feature_profiles() \%>\% ggplot2::ggplot_build()
   b \%>\% autonomics.plot::gg_nrow()
   b \%>\% autonomics.plot::gg_ncol()
}
}
