% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{gg_xlabels}
\alias{gg_xlabels}
\title{Get x labels}
\usage{
gg_xlabels(b)
}
\arguments{
\item{b}{built ggplot (result of ggplot_build)}
}
\value{
x labels
}
\description{
Get x labels from ggplot object
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016 \%>\% magrittr::extract(1:9, )
   b <- object \%>\% autonomics.plot::plot_feature_profiles() \%>\% ggplot2::ggplot_build()
   b \%>\% autonomics.plot::gg_xlabels()
}
}
