% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{is_even}
\alias{is_even}
\alias{is_odd}
\title{Is even/odd?}
\usage{
is_even(x)

is_odd(x)
}
\arguments{
\item{x}{integer}
}
\value{
logical
}
\description{
Is even/odd?
}
\examples{
is_even(13)
is_even(12)
is_odd(13)
is_odd(12)
}
