% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{make_composite_colors}
\alias{make_composite_colors}
\title{Make composite colors}
\usage{
make_composite_colors(svalues)
}
\arguments{
\item{svalues}{character vector}
}
\value{
named character vector (elements = colors, names = color_var levels)
}
\description{
Make composite colors
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   values <- subramanian.2016::metabolon \%>\% autonomics.import::subgroup_values()
   values \%>\% autonomics.plot::make_composite_colors()
}

# GLUTAMINASE
if (require(autonomics.data)){
   values <- autonomics.data::glutaminase \%>\% autonomics.import::subgroup_values()
   values \%>\% autonomics.plot::make_composite_colors()
}
}
