% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{make_volcano_dt}
\alias{make_volcano_dt}
\title{Create volcano datatable}
\usage{
make_volcano_dt(object, ntop = 3)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{ntop}{number of top features (either FC wise or p wise) to be annotated}
}
\value{
data.table
}
\description{
Create volcano datatable
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% make_volcano_dt() \%>\% print()

   object <- subramanian.2016::exiqon
   object \%>\% make_volcano_dt() \%>\% print()
}
}
