% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{order_on_feature_loadings}
\alias{order_on_feature_loadings}
\title{Order on feature loadings}
\usage{
order_on_feature_loadings(object, method, dim, na.impute)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{method}{'pca', 'sma', 'lda', 'pls'}

\item{dim}{dimension}

\item{na.impute}{whether to impute (if applicable)}
}
\value{
re-ordered sumexp
}
\description{
Order on feature loadings
}
