% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions.R
\name{plot}
\alias{plot}
\alias{plot.data.table}
\title{Plot}
\usage{
plot(object, ...)

\method{plot}{data.table}(plotdt, geom, stat, coord_flip = FALSE,
  xlab = NULL, ylab = NULL, title = NULL, facet = NULL, x = NULL,
  y = NULL, group = NULL, ...)
}
\arguments{
\item{object}{data.table}

\item{...}{ggplot2 aesthetic parameters}

\item{geom}{string: ggplot2 geom}

\item{stat}{string: ggplot2 stat}

\item{xlab}{if non-NULL, x axis label}

\item{ylab}{if non-NULL, y axis label}

\item{title}{if non-NULL, title text}

\item{facet}{string: plotdt variable mapped to facet}

\item{horiz}{TRUE or FALSE: whether to flip coordinates}
}
\description{
Plot
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::glutaminase \%>\% prepare_plot_dt()
   plot(object, geom = 'line', stat = 'density', group = sample_id, x = value, color = subgroup, facet = subgroup)
}
}
