% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detects.R
\name{plot_detects_per_subgroup}
\alias{plot_detects_per_subgroup}
\title{Plot detects}
\usage{
plot_detects_per_subgroup(
  object,
  svar = "subgroup",
  color_values = default_color_values(object, svar)
)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{svar}{svar}

\item{color_values}{named character vector (names = colorvar levels, values = color names)}
}
\value{
ggplot object
}
\description{
Plot number of detects, partial detects, and nondetects (or imputes) per subgroup
}
\examples{
require(magrittr)
if (require(autonomics.data)){

   # Before imputation
   object <-  'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
               system.file(package = 'autonomics.data')     \%>\%
               autonomics.import::read_proteingroups()      \%>\%
               autonomics.import::invert(subgroups = c('E_EM', 'E_BM', 'EM_BM'))

   object \%>\% plot_detects_per_subgroup()
   object \%>\% autonomics.import::impute_consistent_nas() \%>\%
              plot_detects_per_subgroup()
}
}
