% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_heatmap.R
\name{plot_feature_heatmap}
\alias{plot_feature_heatmap}
\title{Plot feature heatmap}
\usage{
plot_feature_heatmap(
  object,
  fvar = "feature_id",
  filename = NULL,
  lowColor = "yellow",
  highColor = "blue1",
  midColor = "gray",
  width = 2,
  height = 4,
  fontsize = 0,
  ...
)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{fvar}{string}

\item{filename}{string}

\item{lowColor}{color for downregulation (Default color yellow)}

\item{highColor}{color for upregulation (Default color royal blue)}

\item{midColor}{color for midpoint (Default color gray)}

\item{width}{number: width  in inches (Default width 2 inches)}

\item{height}{number: height in inches (Default height 4 inches)}

\item{fontsize}{number: font size for feature name (Default size 0 for not showing feature names)}

\item{...}{additional parameters handed through to \code{\link[ggplot2]{theme}}}
}
\description{
Plot feature heatmap
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::glutaminase[1:10, ]
   object \%>\% plot_feature_heatmap(fvar='BIOCHEMICAL')
   filename <- tempfile() \%T>\% message()
   object \%>\% plot_feature_heatmap(filename = filename)
}
}
