% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{plot_features_without_printing}
\alias{plot_features_without_printing}
\alias{plot_features}
\alias{plot_feature_bars}
\alias{plot_feature_profiles}
\alias{plot_feature_distributions}
\alias{plot_feature_boxes}
\title{Plot feature profiles / distributions}
\usage{
plot_features_without_printing(object,
  feature_plot = autonomics.plot::default_feature_plots(object)[1],
  x = autonomics.plot::default_x(object, feature_plot[1]),
  color_var = autonomics.plot::default_color_var(object),
  color_values = autonomics.plot::default_color_values(object,
  color_var), shape_var = autonomics.plot::default_shape_var(object),
  group_var = autonomics.plot::default_group_var(object),
  txt_var = autonomics.plot::default_txt_var(object), facet_var = NULL,
  alpha_var = NULL, fvars = autonomics.plot::default_fvars(object),
  title = "", scales = "free_y", x_text_angle = 90, line = FALSE,
  zero_hline = autonomics.plot::default_zero_hline(object),
  xlab = NULL, ylab = NULL, legend.position = "right",
  verbose = FALSE)

plot_features(object, ..., file = NULL, width = NULL, height = NULL)

plot_feature_bars(object, ...)

plot_feature_profiles(object, ...)

plot_feature_distributions(object, ...)

plot_feature_boxes(object, ...)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{feature_plot}{'profiles', 'distributions', 'boxes', 'bars'}

\item{x}{svar mapped to x}

\item{color_var}{svar mapped to color}

\item{color_values}{named color vector (names = color_var levels, values = colors)}

\item{shape_var}{svar mapped to shape (only relevant for profiles)}

\item{group_var}{svar mapped to group}

\item{txt_var}{svar mapped to txt}

\item{facet_var}{svar on which to facet plot}

\item{alpha_var}{svar (logical) mapped to absence/presence of transparancy}

\item{fvars}{fvar used for plot annotation}

\item{title}{plot title}

\item{scales}{'free'}

\item{x_text_angle}{angle of text on x axis}

\item{line}{logical: should line be added?}

\item{zero_hline}{logical: should y=0 line be added?}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{legend.position}{position of legend}

\item{verbose}{logical}

\item{...}{passed to backend}

\item{file}{file to print to. If NULL, ggplot object is returned}

\item{width}{width (inches)}

\item{height}{height (inches)}
}
\description{
Plot feature profiles / distributions
}
\examples{
require(magrittr)
result_dir <- tempdir() \%T>\% message()
if (require(autonomics.data)){
   object <- autonomics.data::billing2016 \%>\% extract(1:10, )
   object \%>\% autonomics.plot::plot_feature_distributions()
   object \%>\% autonomics.plot::plot_feature_profiles()
   object \%>\% autonomics.plot::plot_feature_boxes()

   file <- paste0(result_dir, '/billing2016_distributions.pdf')
   object \%>\% autonomics.plot::plot_feature_distributions(file = file)

   file <- paste0(result_dir, '/billing2016_profiles.pdf')
   object \%>\% autonomics.plot::plot_feature_profiles(file = file)

   file <- paste0(result_dir, '/billing2016_boxes.pdf')
   object \%>\% autonomics.plot::plot_feature_boxes(file = file)
}
if (require(subramanian.2016)){
   subramanian.2016::metabolon[1:10, ] \%>\%
   autonomics.plot::plot_feature_boxes(
      color_var = 'condition', group_var = 'condition', line = TRUE)

   subramanian.2016::exiqon[1:10, ]    \%>\%
   autonomics.plot::plot_feature_boxes(
      color_var = 'condition', group_var = 'condition', line = TRUE)

   subramanian.2016::rnaseq[1:10, ]    \%>\%
   autonomics.plot::plot_feature_boxes(
      color_var = 'condition', group_var = 'condition', line = TRUE)

   result_dir <- tempdir() \%T>\% message()
   file       <- file.path(result_dir, 'feature_boxes.pdf')
   subramanian.2016::metabolon[1:10, ] \%>\%
   autonomics.plot::plot_feature_boxes(
      color_var = 'condition', group_var = 'condition', line = TRUE, file = file)

   \dontrun{
   subramanian.2016::exiqon[1:10, ]    \%>\%
   autonomics.plot::plot_feature_boxes(
      color_var = 'condition', group_var = 'condition', line = TRUE, file = file)
   }

   subramanian.2016::rnaseq[1:10, ]    \%>\%
   autonomics.plot::plot_feature_boxes(
      color_var = 'condition', group_var = 'condition', line = TRUE, file = file)
}
if (require(atkin.2014)){
   object <- atkin.2014::soma \%>\% magrittr::extract(1:10, )
   file <- tempdir() \%>\% paste0('/atkin_hypo_profiles.pdf')
   object \%>\% autonomics.plot::plot_feature_profiles(
                   x = 'time', color_var = 'condition', facet_var = 'subject_id',
                   group_var = 'subject_id', line = TRUE)
   object \%>\% autonomics.plot::plot_feature_profiles(
                   x = 'time', color_var = 'condition', facet_var = 'subject_id',
                   group_var = 'subject_id', line = TRUE, fvar = 'TargetFullName')
   object \%>\% autonomics.plot::plot_feature_profiles(
                   x = 'time', color_var = 'condition', facet_var = 'subject_id',
                   group_var = 'subject_id', line = TRUE, fvar = 'TargetFullName', file = file)
}
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts \%>\% extract(1:10, )
   object \%>\% autonomics.plot::plot_feature_profiles(fvars = 'gene_name')
   object \%>\% autonomics.plot::plot_feature_distributions(fvars = 'gene_name')
   object \%>\% autonomics.plot::plot_feature_boxes(fvars = 'gene_name')

   object <- billing.differentiation.data::protein.ratios \%>\% extract(1:10, )
   object \%>\% autonomics.plot::plot_feature_boxes()
}
if (require(halama.2016)){
   object <- halama.2016::cell.metabolites[1:10, ]
   color_values <- c(Control = 'blue', Vehicle = 'green', `Concentration 1` = 'orange',
                    `Concentration 2` = 'red')
   object$alpha <- object$TREATMENT \%in\% c(0,1)
   object \%>\% autonomics.plot::plot_feature_boxes(
                   x = 'TIME_POINT', color_var = 'GROUP_DESCRIPTION', color_values = color_values,
                   alpha_var = 'alpha')
}

}
\author{
Aditya Bhagwat
}
