% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features_mean_sd.R
\name{plot_features_mean_sd}
\alias{plot_features_mean_sd}
\title{Mean/SD Plot(s)}
\usage{
plot_features_mean_sd(
  object,
  ...,
  exponent = c(0.5, 1, 2),
  complete_data = TRUE,
  facet_labels = NULL,
  facet_grid_labeller = "label_parsed",
  alpha = 0.1
)
}
\arguments{
\item{object}{eSet, EList or SummarizedExperiment}

\item{...}{additonal eSets, ELists, SummarizedExperiments}

\item{exponent}{exponent with which the standard deviation is plotted;
\code{2}: $sigma^2 = 'Varianve'$; \code{1}: $sigma^1 = 'Standard Deviation'$;
\code{0.5}: $sigma^0.5 = sqrt(sigma)$ (default; as used by Law et al (2014));}

\item{complete_data}{\code{\link{logical}} indicating whether to plot only complete data without
\code{\link{NA}} in any replicate;}

\item{facet_labels}{Labels used for \code{object} and \code{...} by \code{\link[ggplot2]{facet_grid}}}

\item{facet_grid_labeller}{Labeller (see \code{\link[ggplot2]{labellers}}) for facetting labels}

\item{alpha}{handed on to \code{\link[ggplot2]{geom_point}}}
}
\value{
\code{ggplot2} object
}
\description{
Mean/variance plots are helpful for the evaluation of homoscedasticity.
}
\examples{
\dontrun{
# Works, but is rather slow
if(require(autonomics.data)){
   require(magrittr)

  # Simple plot
     plot_features_mean_sd(autonomics.data::ALL)

  # Compare with the loged version of the data
     logALL <- autonomics.data::ALL
     autonomics.import::exprs(logALL) \%<>\% log2()
     plot_features_mean_sd(logALL)

  # Combined plot
     plot_features_mean_sd(autonomics.data::ALL, logALL)

  # Same thing, but with descriptive facet labels
     plot_features_mean_sd(
        autonomics.data::ALL,
        logALL,
        facet_labels = c("'Raw Data'", "'Logarithmized Data'~(log[2])"))
}
}
}
\references{
Law, C.W., Chen, Y., Shi, W., and Smyth, G.K. (2014). voom: precision
weights unlock linear model analysis tools for RNA-seq read counts. Genome Biology 15, R29.
}
