% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ora.R
\name{plot_ora}
\alias{plot_ora}
\title{Plot ORA results}
\usage{
plot_ora(ora_results, top_n = 5, collections = c("gobp", "gocc",
  "gomf", "kegg"), directions = c("neg", "pos", "both"),
  contrasts = NULL, facet_var_x = "contrast",
  facet_var_y = "direction", trim_pathway_id = FALSE)
}
\arguments{
\item{ora_results}{\code{\link{data.frame}} as (silently) returned from \code{autonomics.ora::run_ora_on_eset}}

\item{top_n}{number of factors chosen for plotting (by rank)}

\item{collections}{(sub)ontologies/collections to plot}

\item{directions}{what coefficient/effect direction of a given contrast the set tested is chosen from}

\item{contrasts}{contrasts to filter on}

\item{facet_var_x}{variable name to \code{\link[ggplot2]{facet_grid}} on}

\item{facet_var_y}{variable name to \code{\link[ggplot2]{facet_grid}} on}

\item{trim_pathway_id}{whether to trim the 'GOxxxxx:' from labels on the plot}
}
\value{
ggplot2 object
}
\description{
Plot ORA results
}
\author{
Aditya Bhagwat, Johannes Graumann
}
