% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlayed_sample_distributions.R
\name{plot_overlayed_sample_distributions}
\alias{plot_overlayed_sample_distributions}
\title{Plot overlayed sample distributions}
\usage{
plot_overlayed_sample_distributions(object,
  color_var = autonomics.plot::default_color_var(object),
  color_values = autonomics.plot::default_color_values(object,
  color_var))
}
\arguments{
\item{object}{SummarizedExperiment}

\item{facet_var}{svar to facet on}
}
\description{
Plot overlayed sample distributions
}
\examples{
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   facet_var <- 'subgroup'
   object \%>\% autonomics.plot::plot_overlayed_sample_distributions()
}
}
