% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_projected_samples}
\alias{plot_projected_samples}
\alias{plot_pca_samples}
\alias{plot_sma_samples}
\alias{plot_lda_samples}
\alias{plot_pls_samples}
\title{Plot PCA/LDA/PLS/SMA sample scores}
\usage{
plot_projected_samples(
  object,
  ...,
  listed_objects = NULL,
  method = c("pca", "lda", "sma", "pls")[1],
  implementation = c("mixOmics::plsda", "mixOmics::splsda", "ropls::opls")[1],
  dims = 1:2,
  color_var = "subgroup",
  color_values = default_color_values(object, color_var),
  shape_var = default_shape_var(object),
  size_var = NULL,
  txt_var = default_txt_var(object),
  txt_outliers = FALSE,
  group_var = NULL,
  split_var = NULL,
  facet_var = NULL,
  scales = ifelse(is.null(facet_var), "fixed", "free"),
  labs = list(color = stringi::stri_trans_totitle(color_var), shape =
    stringi::stri_trans_totitle(shape_var)),
  nrow = NULL,
  mention_method = ifelse(is.null(facet_var), TRUE, FALSE),
  title = NULL,
  na.impute = FALSE,
  legend.position = "right",
  plot = TRUE
)

plot_pca_samples(
  object,
  ...,
  listed_objects = NULL,
  dims = 1:2,
  color_var = "subgroup",
  color_values = default_color_values(object, color_var),
  shape_var = default_shape_var(object),
  size_var = NULL,
  txt_var = default_txt_var(object),
  txt_outliers = FALSE,
  group_var = NULL,
  split_var = NULL,
  facet_var = NULL,
  scales = ifelse(is.null(facet_var), "fixed", "free"),
  labs = list(color = stringi::stri_trans_totitle(color_var), shape =
    stringi::stri_trans_totitle(shape_var)),
  nrow = NULL,
  mention_method = ifelse(is.null(facet_var), TRUE, FALSE),
  title = NULL,
  na.impute = FALSE,
  legend.position = "right",
  plot = TRUE
)

plot_sma_samples(
  object,
  ...,
  listed_objects = NULL,
  dims = 1:2,
  color_var = "subgroup",
  color_values = default_color_values(object, color_var),
  shape_var = default_shape_var(object),
  size_var = NULL,
  txt_var = default_txt_var(object),
  txt_outliers = FALSE,
  group_var = NULL,
  split_var = NULL,
  facet_var = NULL,
  scales = ifelse(is.null(facet_var), "fixed", "free"),
  labs = list(color = stringi::stri_trans_totitle(color_var), shape =
    stringi::stri_trans_totitle(shape_var)),
  nrow = NULL,
  mention_method = ifelse(is.null(facet_var), TRUE, FALSE),
  title = NULL,
  na.impute = FALSE,
  legend.position = "right",
  plot = TRUE
)

plot_lda_samples(
  object,
  ...,
  listed_objects = NULL,
  dims = 1:2,
  color_var = "subgroup",
  color_values = default_color_values(object, color_var),
  shape_var = default_shape_var(object),
  size_var = NULL,
  txt_var = default_txt_var(object),
  txt_outliers = FALSE,
  group_var = NULL,
  split_var = NULL,
  facet_var = NULL,
  scales = ifelse(is.null(facet_var), "fixed", "free"),
  labs = list(color = stringi::stri_trans_totitle(color_var), shape =
    stringi::stri_trans_totitle(shape_var)),
  nrow = NULL,
  mention_method = ifelse(is.null(facet_var), TRUE, FALSE),
  title = NULL,
  na.impute = FALSE,
  legend.position = "right",
  plot = TRUE
)

plot_pls_samples(
  object,
  ...,
  listed_objects = NULL,
  implementation = c("mixOmics::plsda", "mixOmics::splsda", "ropls::opls")[1],
  dims = 1:2,
  color_var = "subgroup",
  color_values = default_color_values(object, color_var),
  shape_var = default_shape_var(object),
  size_var = NULL,
  txt_var = default_txt_var(object),
  txt_outliers = FALSE,
  group_var = NULL,
  split_var = NULL,
  facet_var = NULL,
  scales = ifelse(is.null(facet_var), "fixed", "free"),
  labs = list(color = stringi::stri_trans_totitle(color_var), shape =
    stringi::stri_trans_totitle(shape_var)),
  nrow = NULL,
  mention_method = ifelse(is.null(facet_var), TRUE, FALSE),
  title = NULL,
  na.impute = FALSE,
  legend.position = "right",
  plot = TRUE
)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{...}{Additional SummarizedExperiment(s)}

\item{listed_objects}{Additional SummarizedExperiment(s), wrapped in a \code{link{list}}}

\item{method}{'pca', 'lda', 'pls', 'sma'}

\item{implementation}{which implementation of pls to use (see \code{\link{project}})}

\item{dims}{dimensions}

\item{color_var}{svar mapped to color}

\item{color_values}{color values vector}

\item{shape_var}{svar mapped to shape}

\item{size_var}{svar mapped to size}

\item{txt_var}{svar mapped to txt}

\item{txt_outliers}{logical(1): whether to annotate outliers}

\item{group_var}{svar mapped to group}

\item{split_var}{svar on which to split data prior to transformation}

\item{facet_var}{svar on which to facet sample biplot}

\item{scales}{'free' or 'fixed'}

\item{labs}{named \code{\link{list}} handed on to \code{\link[ggplot2]{labs}}}

\item{nrow}{integer}

\item{mention_method}{Whether to use \code{method} in title and/or facet labels}

\item{title}{title}

\item{na.impute}{TRUE or FALSE}

\item{legend.position}{character}

\item{plot}{TRUE (default) or FALSE}
}
\description{
Plot PCA/LDA/PLS/SMA sample scores
}
\examples{
if (require(autonomics.data)){
   require(magrittr)

   # GLUTAMINASE
      object <- autonomics.data::glutaminase
      object \%>\% plot_pca_samples()
      object \%>\% plot_lda_samples()
      object \%>\% plot_pls_samples()
      object \%>\% plot_projected_samples(
                    method = c('pca', 'lda', 'sma', 'pls'),
                    facet_var = c('PCA', 'LDA', 'SMA', 'PLS'),
                    nrow = 2)

   # STEM CELL COMPARISON
      object <- autonomics.data::stemcomp.proteinratios
      object \%>\% plot_pca_samples()
      object \%>\% plot_pca_samples(dims = c(3,4))

      object \%>\% plot_projected_samples(
                    method = c('pca', 'lda', 'sma', 'pls'),
                    facet_var = c('OK PCA', 'Not so nice LDA', 'Nice SMA', 'Very nice PLS'),
                    nrow = 2)

}

}
\author{
Aditya Bhagwat, Johannes Graumann
}
