% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_projected_samples_and_features}
\alias{plot_projected_samples_and_features}
\alias{plot_pca_samples_and_features}
\alias{plot_sma_samples_and_features}
\alias{plot_lda_samples_and_features}
\alias{plot_pls_samples_and_features}
\title{Plot pca samples and features}
\usage{
plot_projected_samples_and_features(
  object,
  method,
  implementation = NULL,
  dims = c(1, 2),
  na.impute = FALSE,
  geom = default_feature_plots(object)[1],
  result_dir = NULL,
  x = default_x(object, geom),
  color_var = default_color_var(object),
  color_values = default_color_values(object, color_var),
  shape_var = NULL,
  group_var = default_group_var(object),
  txt_var = default_txt_var(object),
  line = default_line(object),
  n = 9
)

plot_pca_samples_and_features(object, ...)

plot_sma_samples_and_features(object, ...)

plot_lda_samples_and_features(object, ...)

plot_pls_samples_and_features(object, ...)
}
\arguments{
\item{object}{eset}

\item{method}{'pca', 'lda', or 'pls'}

\item{implementation}{NULL or "package::function"}

\item{dims}{pc dimensions to plot}

\item{na.impute}{TRUE or FALSE}

\item{geom}{value in \code{\link[autonomics.plot]{FEATURE_PLOTS}}}

\item{result_dir}{NULL or result directory path}

\item{x}{svar mapped to x in feature plots}

\item{color_var}{svar mapped to color in sample and feature plots}

\item{color_values}{color vector (names = color_var levels, values = colors)}

\item{shape_var}{svar mapped to shape in sample and feature plots}

\item{group_var}{svar mapped to group in feature plots}

\item{txt_var}{svar mapped to txt in feature plots}

\item{line}{whether to connect points in feature plot with line (logical)}

\item{n}{no of features to be plotted}

\item{...}{passed to plot_projected_samples_and_features}
}
\description{
Plots pca sample and features results in a combined plot
}
\examples{
\dontrun{
   if (require(autonomics.data)){

      # STEM CELL COMPARISON
        require(magrittr)
        result_dir <- tempdir() \%T>\% message()
        object <- autonomics.data::stemcomp.proteinratios
        object \%>\% plot_pca_samples_and_features()
        object \%>\% plot_pca_samples_and_features(geom = 'violin')
        object \%>\% plot_pca_samples_and_features(geom = 'violin', na.impute = TRUE)
        object \%>\% plot_pca_samples_and_features(result_dir = result_dir)
        object \%>\% plot_pca_samples_and_features(result_dir = result_dir, geom = 'bar')

     # GLUTAMINASE
        object <- autonomics.data::glutaminase
        object \%>\% plot_pca_samples_and_features(n=2)

   }
}
}
