% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_features.R
\name{plot_projection_features}
\alias{plot_projection_features}
\alias{plot_pca_features}
\alias{plot_sma_features}
\alias{plot_lda_features}
\alias{plot_pls_features}
\title{Plot PCA/PLS/LDA/SMA features}
\usage{
plot_projection_features(object, method, implementation = NULL,
  geom = default_feature_plots(object)[1],
  fvars = default_fvars(object), dim = 1, n = 9, na.impute = FALSE,
  title = sprintf("X\%d", dim), file = NULL, ...)

plot_pca_features(object, implementation = NULL,
  geom = default_feature_plots(object)[1],
  fvars = default_fvars(object), dim = 1, n = 9, na.impute = FALSE,
  title = sprintf("X\%d", dim), file = NULL, ...)

plot_sma_features(object, implementation = NULL,
  geom = default_feature_plots(object)[1],
  fvars = default_fvars(object), dim = 1, n = 9, na.impute = FALSE,
  title = sprintf("X\%d", dim), file = NULL, ...)

plot_lda_features(object, implementation = NULL,
  geom = default_feature_plots(object)[1],
  fvars = default_fvars(object), dim = 1, n = 9, na.impute = FALSE,
  title = sprintf("X\%d", dim), file = NULL, ...)

plot_pls_features(object, implementation = NULL,
  geom = default_feature_plots(object)[1],
  fvars = default_fvars(object), dim = 1, n = 9, na.impute = FALSE,
  title = sprintf("X\%d", dim), file = NULL, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{method}{'pca', 'lda', 'pls', 'sma'}

\item{implementation}{'character' or NULL}

\item{geom}{value in \code{\link[autonomics.plot]{FEATURE_PLOTS}}}

\item{fvars}{fvars used for plot annotation}

\item{dim}{principal component dimension}

\item{n}{number of top features to plot}

\item{na.impute}{TRUE or FALSE}

\item{title}{title}

\item{file}{file}

\item{...}{passed to \code{\link[autonomics.plot]{plot_features}}}
}
\description{
Plots top PCA/PLS/LDA/SMA features.
Uses factor loadings in object when available.
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% plot_pca_features(n=9)
   object \%>\% plot_pca_features(geom = 'bar')
}

# GLUTAMINASE
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% plot_pca_features()
}

}
