% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions.R
\name{plot_sample_boxplots}
\alias{plot_sample_boxplots}
\alias{plot_sample_boxplots.matrix}
\alias{plot_sample_boxplots.SummarizedExperiment}
\title{Plot sample boxplots}
\usage{
plot_sample_boxplots(object, ...)

\method{plot_sample_boxplots}{matrix}(object, sdata, coord_flip = TRUE,
  xlab = "", title = "Sample boxplots", ...)

\method{plot_sample_boxplots}{SummarizedExperiment}(object,
  coord_flip = TRUE, xlab = "", title = "Sample boxplots", ...)
}
\arguments{
\item{object}{matrix or SummarizedExperiment}

\item{...}{ggplot2 parameters or aesthetics, passed to plot.data.table}

\item{sdata}{data.table}

\item{horiz}{TRUE or FALSE}
}
\value{
ggplot2 object
}
\description{
Plot sample boxplots
}
\examples{
if (require(autonomics.data)){

   # matrix
   require(magrittr)
   object <- stemcomp.proteinratios \%>\% autonomics.import::exprs()
   sdata  <- stemcomp.proteinratios \%>\% autonomics.import::sdata()
   plot_sample_boxplots(object, sdata)
   plot_sample_boxplots(object, sdata, color = subgroup)

}
}
