% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions.R
\name{plot_sample_densities}
\alias{plot_sample_densities}
\alias{plot_sample_densities.matrix}
\alias{plot_sample_densities.SummarizedExperiment}
\alias{plot_sample_densities.list}
\alias{plot_sample_violins.matrix}
\alias{plot_sample_violins.SummarizedExperiment}
\title{Plot sample densities}
\usage{
plot_sample_densities(object, ...)

\method{plot_sample_densities}{matrix}(object, sdata, color = if
  ("subgroup" \%in\% names(sdata)) "subgroup" else "sample_id", ...,
  facet = NULL, xlab = NULL, ylab = NULL, title = NULL)

\method{plot_sample_densities}{SummarizedExperiment}(object, color = if
  ("subgroup" \%in\% autonomics.import::svars(object)) "subgroup" else
  "sample_id", ..., facet = NULL)

\method{plot_sample_densities}{list}(object, color = "sample_id", ...,
  facet = NULL)

\method{plot_sample_violins}{matrix}(object, horiz = TRUE,
  color_var = "sample_id")

\method{plot_sample_violins}{SummarizedExperiment}(object, x = NULL,
  color_var = default_color_var(object),
  color_values = default_color_values(object, color_var),
  facet_var = NULL, displayed_features = NULL, ylab = "",
  xlab = "",
  title = sprintf("Per sample distribution of all \%d features",
  nrow(object)), file = NULL)
}
\arguments{
\item{object}{matrix, SummarizedExperiment, or list of SummarizedExperiments}

\item{...}{passed to ggplot2::aes()}

\item{sdata}{dataframe}

\item{color}{string (varname). For a list object, an equidimensional stringvector (with varlevels) is also allowed.}

\item{facet}{string (varname). For a list object, an equidimensional stringvector (with varlevels) is also allowed.}
}
\value{
ggplot2 object
}
\description{
Plot sample densities
}
\examples{
if (require(autonomics.data)){

   # STEM CELL COMPARISON
   require(magrittr)
   plot_sample_densities(stemcomp.proteinratios)
   plot_sample_densities(object = autonomics.import::exprs(stemcomp.proteinratios),
                         sdata  = autonomics.import::sdata(stemcomp.proteinratios))

   # GLUTAMINASE
   plot_sample_densities(glutaminase, color = 'EXPERIMENT', facet = c('CONCENTRATION', 'TIME_POINT'))
   plot_sample_densities(object = autonomics.import::exprs(glutaminase),
                         sdata  = autonomics.import::sdata(glutaminase))
   plot_sample_densities(object = list(glutaminase \%>\% magrittr::extract(, 1:12),
                                       glutaminase \%>\% magrittr::extract(, 13:ncol(.))),
                         color = c('UT_h10', 'other'),
                         facet = c('UT_h10', 'other'))
}
}
