% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions.R
\name{plot_sample_distributions}
\alias{plot_sample_distributions}
\title{Plot distribution across features for each sample}
\usage{
plot_sample_distributions(object, x = NULL,
  color_var = default_color_var(object),
  color_values = default_color_values(object, color_var),
  facet_var = NULL, displayed_features = NULL, ylab = "",
  xlab = "",
  title = sprintf("Per sample distribution of all \%d features",
  nrow(object)), file = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{x}{string: sample variable mapped to x axis}

\item{color_var}{string: sample variable mapped to color}

\item{color_values}{string vector: names = subgroups, values = colors}

\item{facet_var}{string: sample variable for faceting}

\item{displayed_features}{features to be displayed in the sample distributions (vector of numeric indexes or character \code{feature_id}s)}

\item{ylab}{label of y axis (character)}

\item{xlab}{label of y axis (character)}

\item{title}{title (character)}

\item{file}{file to which to print plot}
}
\value{
ggplot2 object
}
\description{
The distributions are plotted as a combination of a violin plot and a box plot.
A separate violin is created for each value of 'x' (rows) and 'facet_var (column).
The box is colored per value of color_var
}
\examples{
if (require(autonomics.data)){

   # ALL
      require(magrittr)
      object <- autonomics.data::ALL[, 1:30]
      object \%>\% plot_sample_distributions(x = 'cod', color_var = 'BT')

      # faceting works, but is not useful here, as there is no block variable
      object \%>\% plot_sample_distributions(x = 'cod', facet_var = 'sex', color_var = 'BT')

   # STEM CELL COMPARISON
      object <- autonomics.data::stemcomp.proteinratios
      object \%>\% plot_sample_distributions()

   # GLUTMINASE
      object <- autonomics.data::glutaminase
      object \%>\% plot_sample_distributions2()
}
}
\author{
Aditya Bhagwat
}
