% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions2.R
\name{plot_sample_distributions2}
\alias{plot_sample_distributions2}
\title{Plot distribution across features for each sample}
\usage{
plot_sample_distributions2(object, ..., x = NULL,
  color_var = autonomics.plot::default_color_var(object),
  color_values = autonomics.plot::default_color_values(object, color_var),
  displayed_features = NULL, jitter_features = FALSE, horizontal = TRUE,
  xlab = "", ylab = "",
  title = sprintf("Per sample distribution of all \%d features",
  nrow(object)))
}
\arguments{
\item{object}{eSet}

\item{...}{additonal eSets, ELists, SummarizedExperiments or matrices AND further parameters for \code{\link{core_ggplot_from_data}}}

\item{x}{sample variable mapped to x axis}

\item{color_var}{sample variable mapped to color}

\item{color_values}{color values vector (names = subgroups, values = colors)}

\item{displayed_features}{features to be displayed in the sample distributions (vector of numeric indexes or character \code{feature_id}s)}

\item{jitter_features}{whether to use geom_jitter or not for the displayed features}

\item{horizontal}{whether to lay the plot out horizontally (measurements on the x, sample names on the y axis)}

\item{xlab}{label of y axis (character)}

\item{ylab}{label of y axis (character)}

\item{title}{title (character)}
}
\value{
ggplot2 object
}
\description{
The distributions are plotted as a combination of a violin plot and a box plot.
A separate violin is created for each value of 'x' (rows) and 'facet_var (column).
The box is colored per value of color_var
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   # Simple plot
     autonomics.plot::plot_sample_distributions2(autonomics.data::stemcomp.proteinratios)
   # Same thing vertical
     plot_sample_distributions2(autonomics.data::stemcomp.proteinratios, horizontal = FALSE)
   # More complex using facetting
     plot_sample_distributions2(
        object = autonomics.data::stemcomp.proteinratios,
        2^autonomics.import::exprs(autonomics.data::stemcomp.proteinratios),
        facet2_var = c('Logarithmized', 'Raw'))
   # Yet more crazy explicitly marking the feature with the maximum value
     object <- autonomics.data::stemcomp.proteinratios
     plot_sample_distributions2(
        object,
        2^autonomics.import::exprs(object),
        facet2_var = c('Logarithmized', 'Raw'),
        displayed_features =
           autonomics.import::exprs(object) \%>\%
           magrittr::equals(autonomics.import::exprs(object) \%>\%
                            max(na.rm = TRUE)) \%>\%
           which(arr.ind = TRUE) \%>\%
           magrittr::extract(,'row'))
}
}
\seealso{
\code{\link{core_ggplot_from_data}}
}
\author{
Aditya Bhagwat, Johannes Graumann
}
