% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_sample_distrs_n_pca}
\alias{plot_sample_distrs_n_pca}
\title{Plot sample distributions and pca}
\usage{
plot_sample_distrs_n_pca(object, descr = "",
  color_var = default_color_var(object),
  color_values = default_color_values(object),
  shape_var = default_shape_var(object),
  txt_var = default_txt_var(object), displayed_features = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{descr}{string: used as plot title and plot name}

\item{color_var}{string: svar mapped to color}

\item{color_values}{named string vector: names = color_var levels, values = colors}

\item{shape_var}{string: svar mapped to shape}

\item{txt_var}{string: svar mapped to txt}

\item{displayed_features}{number or string vector: features to be displayed in the sample distributions)}
}
\value{
list(descr.distr = plot1, descr.pca = plot2)
}
\description{
Plot sample distributions and pca
}
\examples{
if (require(autonomics.data)){
   # GLUTAMINASE
   require(magrittr)
   object <- autonomics.data::glutaminase
   object \%>\% plot_sample_distrs_n_pca()
}
}
