% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions.R
\name{plot_sample_violins}
\alias{plot_sample_violins}
\alias{plot_sample_violins.matrix}
\alias{plot_sample_violins.SummarizedExperiment}
\title{Plot sample violins}
\usage{
plot_sample_violins(object, ...)

\method{plot_sample_violins}{matrix}(object, sdata, coord_flip = TRUE,
  xlab = "", title = "Sample violins", ...)

\method{plot_sample_violins}{SummarizedExperiment}(object,
  coord_flip = TRUE, xlab = "", title = "Sample violins", ...)
}
\arguments{
\item{object}{matrix or SummarizedExperiment}

\item{...}{ggplot parameters or aesthetics}

\item{sdata}{dataframe}
}
\value{
ggplot2 object
}
\description{
Plot sample violins
}
\examples{
if (require(autonomics.data)){
   # matrix
      require(magrittr)
      object <- stemcomp.proteinratios \%>\% autonomics.import::exprs()
      sdata  <- stemcomp.proteinratios \%>\% autonomics.import::sdata()
      plot_sample_violins(object, sdata)
      plot_sample_violins(object, sdata, fill = subgroup)
      plot_sample_violins(object, sdata, fill = subgroup, horiz = FALSE)

   # SummarizedExperiment
      object <- stemcomp.proteinratios
      plot_sample_violins(object,fill=subgroup)
      plot_sample_violins(object,fill=subgroup, horiz=FALSE)
}
}
