% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venns.R
\name{plot_venn}
\alias{plot_venn}
\alias{plot_venns}
\title{Plot Venn/Euler diagram}
\usage{
plot_venn(x, color_values = if (assertive.properties::has_names(x)) {    
  autonomics.plot::make_colors(names(x)) } else {    
  autonomics.plot::make_gg_colors(1:length(x)) }, euler = FALSE,
  filename = NULL, width = 7, height = width, title = NULL,
  scaled = FALSE, cat.dist = 0.3, margin = cat.dist,
  main.cex = 1.5, main.pos = c(0.5, 0.97), ...)

plot_venns(...)
}
\arguments{
\item{x}{named list of character vectors}

\item{color_values}{character vector}

\item{euler}{logical(1): whether to use eulerr::euler() rather than VennDiagram::venn.diagram()}

\item{filename}{character vector}

\item{width}{width  in inches}

\item{height}{height in inches}

\item{title}{figure title}

\item{scaled}{logical: whether or not to scale}

\item{cat.dist}{numeric(1): category location (as distance from circle edge)}

\item{main.cex}{numeric(1): scaling factor for title size}

\item{main.pos}{numeric(2): title location}

\item{...}{passed to either VennDiagram::venn.diagram(...) or eulerr::plot.euler(...)}
}
\description{
Plot Venn/Euler diagram
}
\examples{
x <- list(A = c('apple', 'pear'), B = c('pear', 'orange'))
filename <- NULL
autonomics.plot::plot_venn(x)
autonomics.plot::plot_venn(x, euler = TRUE)
filename <- tempfile()
autonomics.plot::plot_venn(x, filename = filename)

}
