% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot volcano}
\usage{
plot_volcano(object, ntop = 3, nrow = NULL, legend_position = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{ntop}{number: n top features to be annotated}

\item{nrow}{number: n rows in faceted plot}

\item{legend_position}{handed to \code{\link[ggplot2]{theme}} as 'legend.position'}
}
\value{
ggplot object
}
\description{
Plot volcano
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% plot_volcano()
}
}
