% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pls}
\alias{pls}
\title{Partial least squares analysis (PLS)}
\usage{
pls(object, implementation = NULL, ndim = 2, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{implementation}{'mixOmics::plsda', 'mixOmics::splsda', or 'ropls::opls'}

\item{ndim}{Number of latent variables to include in results}

\item{...}{only inlcuded to keep \code{\link[autonomics.plot]{project}} generic}
}
\value{
\preformatted{list(samples  =   sample scores,       # matrix:  nsample x ndim
     features =  feature loadings,     # matrix: nfeature x ndim
     var      = variance percentages   # vector:        1 x ndim
}
}
\description{
Returns \code{PLS} sample scores, feature loadings, and latent variable variances.
}
\details{
\code{PLS} is performed using the (\code{NA}-friendly) \code{NIPALS} algorithm.
}
\examples{
if (require(autonomics.data)){
   require(magrittr)

   # STEM CELL DIFFERENTIATION
   object <- 'extdata/stemdiff/maxquant/proteinGroups.txt' \%>\%
              system.file(package = 'autonomics.data') \%>\%
              autonomics.import::read_proteingroups()
   object \%>\% autonomics.plot::pls() \%>\% str()

  # GLUTAMINASE
   object <- 'extdata/glutaminase/glutaminase.xlsx' \%>\%
              system.file(package='autonomics.data') \%>\%
              autonomics.import::read_metabolon()
   object \%>\% autonomics.plot::pls() \%>\% str()
   \dontrun{ # slow
      object \%>\% autonomics.plot::pls(implementation = 'ropls::opls') \%>\% str()
   }
}
}
\seealso{
\code{\link[autonomics.plot]{pca}},
\code{\link[autonomics.plot]{sma}},
\code{\link[autonomics.plot]{lda}}
}
\author{
Aditya Bhagwat
}
