% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions.R
\name{prepare_plot_dt}
\alias{prepare_plot_dt}
\title{Prepare plot datatable}
\usage{
prepare_plot_dt(exprsmat, sdata)
}
\arguments{
\item{exprsmat}{exprs matrix (feature x sample)}

\item{sdata}{sample dataframe}
}
\value{
datatable
}
\description{
Prepare plot datatable
}
\examples{
if (require(autonomics.data)){

   # STEM CELL COMPARISON
   require(magrittr)
   prepare_plot_dt(autonomics.import::exprs(stemcomp.proteinratios),
                   autonomics.import::sdata(stemcomp.proteinratios))

   # GLUTAMINASE
   prepare_plot_dt(autonomics.import::exprs(glutaminase),
                   autonomics.import::sdata(glutaminase))

}
}
