% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions.R
\name{print_sample_distributions}
\alias{print_sample_distributions}
\title{Print sample distributions}
\usage{
print_sample_distributions(ggplot_obj, file)
}
\arguments{
\item{ggplot_obj}{ggplot object}

\item{file}{file to print to}
}
\description{
Print sample distributions
}
