% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{project}
\alias{project}
\title{Project}
\usage{
project(object, method = "pca", na.impute = FALSE,
  implementation = NULL, ndim = 2)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{method}{'pca', 'sma', 'lda', 'pls'}

\item{na.impute}{TRUE or FALSE}

\item{implementation}{which implementation of the method to use}

\item{ndim}{Number of 'dimensions' (e.g. PCA: dimensions == components) to return}
}
\value{
list(samples, features, var)
}
\description{
Return sample scores, feature loadings, and component variances.
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::glutaminase
   object \%>\% autonomics.plot::project() \%>\% str()
}
}
\seealso{
\code{\link[autonomics.plot]{pca}},
         \code{\link[autonomics.plot]{sma}},
         \code{\link[autonomics.plot]{lda}},
         \code{\link[autonomics.plot]{pls}}
}
\author{
Aditya Bhagwat
}
