% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{sma}
\alias{sma}
\title{Spectral Map Analysis (SMA)}
\usage{
sma(object, na.impute = FALSE, ndim = 2, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{na.impute}{whether to \code{\link[autonomics.preprocess]{impute}} prior to \code{SMA}}

\item{ndim}{Number of SMA components to include in results}

\item{...}{used to keep \code{\link{project}} generic}
}
\value{
\preformatted{list(samples  =   sample scores,       # matrix:  nsample x ndim
     features =  feature loadings,     # matrix: nfeature x ndim
     var      = variance percentages   # vector:        1 x ndim
}
}
\description{
Returns \code{SMA} sample scores, feature loadings, and component variances.
}
\details{
\code{SMA} is based on the (\code{NA}-unfriendly) \code{SVD}, so only \code{NA}-free features are used. \cr
\code{SMA} and \code{PCA} are related: their \code{SVD} is identical, but score and loading computation differ:\preformatted{   PCA Scores  : U * D
   SMA Scores  : U * D^(1/2)

PCA Loadings:         V
   SMA Loadings: D^(1/2)*V
}
}
\examples{
require(magrittr)
if (require(autonomics.data)){

   # STEM CELL DIFFERENTIATION
      require(magrittr)
      object <- 'extdata/stemdiff/maxquant/proteinGroups.txt' \%>\%
                 system.file(package='autonomics.data')       \%>\%
                 autonomics.import::read_proteingroups()
      object \%>\% sma() \%>\% str()

   # GLUTAMINASE
      object <- 'extdata/glutaminase/glutaminase.xlsx'    \%>\%
                 system.file(package = 'autonomics.data') \%>\%
                 autonomics.import::read_metabolon()
      object \%>\% sma() \%>\% str()
}
}
\references{
Wouters et al (2003) Graphical exploration of gene expression data: a comparative study of
three multivariate methods.
}
\seealso{
\code{\link[autonomics.plot]{pca}},
\code{\link[autonomics.plot]{lda}},
\code{\link[autonomics.plot]{pls}}
}
\author{
Aditya Bhagwat, Laure Cougnaud
}
