% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validify.R
\name{validify_shape_values}
\alias{validify_shape_values}
\title{Validify shape_var values}
\usage{
validify_shape_values(object, shape_var)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{shape_var}{svar mapped to shape}
}
\description{
Validify shape_var values
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   autonomics.data::billing2016 \%>\%
   autonomics.plot::validify_shape_values('replicate')
}
}
