% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{write_projected_features}
\alias{write_projected_features}
\title{Write pca/lda/pls/sma features}
\usage{
write_projected_features(object, result_dir, method)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{result_dir}{string: result dir}

\item{method}{string: 'pca', 'lda', 'pls', 'sma'}
}
\description{
Write pca/lda/pls/sma features
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   result_dir <- tempdir() \%T>\% message()
   object \%>\% add_pca() \%>\%
              write_projected_features(method = 'pca', result_dir = result_dir)
}
}
