% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{compute_feature_plot_dims}
\alias{compute_feature_plot_dims}
\title{Compute appropriate feature plot dims}
\usage{
compute_feature_plot_dims(p)
}
\arguments{
\item{p}{ggplot object}
}
\description{
Compute appropriate feature plot dims
}
\examples{
require(magrittr)
require(SummarizedExperiment)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios \%>\% magrittr::extract(1:9, )
   object \%>\% autonomics.plot::plot_feature_profiles()      \%>\%
              autonomics.plot::compute_feature_plot_dims()
   object \%>\% autonomics.plot::plot_feature_distributions() \%>\%
              autonomics.plot::compute_feature_plot_dims()
   object \%>\% autonomics.plot::plot_feature_boxes()         \%>\%
              autonomics.plot::compute_feature_plot_dims()
}
if (require(atkin.2014)){
   object <- atkin.2014::soma \%>\% magrittr::extract(1:10, )
   p <- object \%>\% autonomics.plot::plot_feature_profiles(
                        x = 'time', color_var = 'condition', facet_var = 'subject_id',
                        line = TRUE, fvars = 'TargetFullName')
   p \%>\% autonomics.plot::compute_feature_plot_dims()

}
}
