% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_feature_plot_df.R
\name{create_feature_plot_df}
\alias{create_feature_plot_df}
\title{Create dataframe for feature plots}
\usage{
create_feature_plot_df(object, fvars, verbose = FALSE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{fvars}{fvars used to annotate each plot}

\item{verbose}{logical}
}
\value{
dataframe
}
\description{
Create dataframe for feature plots
}
\examples{
if (require(autonomics.data)){
  require(magrittr)

  # STEM CELL COMPARISON (Max Quant)
  autonomics.data::stemcomp.proteinratios \%>\%
  create_feature_plot_df(c('feature_id', 'Gene names')) \%>\% head()

  # GLUTAMINASE
  autonomics.data::glutaminase \%>\%
  create_feature_plot_df(c('feature_id', 'BIOCHEMICAL')) \%>\% head()


}
}
\author{
Aditya Bhagwat
}
