% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_feature_plot_df.R
\name{create_feature_plot_df}
\alias{create_feature_plot_df}
\title{Create dataframe for feature plots}
\usage{
create_feature_plot_df(object, fvars, verbose = FALSE)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{fvars}{fvars used to annotate each plot}

\item{verbose}{logical}
}
\value{
dataframe
}
\description{
Create dataframe for feature plots
}
\examples{
require(magrittr)

# STEM CELL COMPARISON (Max Quant)
if (require(autonomics.data)){
   autonomics.data::stemcomp.proteinratios \%>\%
   autonomics.plot::create_feature_plot_df(c('feature_id', 'Gene names')) \%>\% head()
}

# STEM CELL DIFFERENTIATION (Max Quant)
if (require(autonomics.data)){
   autonomics.data::stemdiff.proteinratios \%>\%
   autonomics.plot::create_feature_plot_df('Gene names') \%>\% head()
}
if (require(billing.differentiation.data)){
   billing.differentiation.data::rna.voomcounts \%>\%
      autonomics.plot::create_feature_plot_df('gene_name') \%>\% head()
   billing.differentiation.data::protein.ratios \%>\%
      autonomics.plot::create_feature_plot_df('Gene names') \%>\% head()
}
}
\author{
Aditya Bhagwat
}
