% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_feature_plots}
\alias{default_feature_plots}
\title{default feature_plots}
\usage{
default_feature_plots(object)
}
\arguments{
\item{object}{eset}
}
\value{
default value of feature_plots
}
\description{
default feature_plots
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   autonomics.data::billing2016 \%>\% default_feature_plots()
}
if (require(billing.differentiation.data)){
   billing.differentiation.data::protein.ratios \%>\%
      default_feature_plots()
   billing.differentiation.data::rna.voomcounts \%>\%
      default_feature_plots()
}
if (require(atkin.2014)){
   atkin.2014::soma \%>\% autonomics.plot::default_feature_plots()
}
}
