% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_feature_plots}
\alias{default_feature_plots}
\title{default feature_plots}
\usage{
default_feature_plots(object)
}
\arguments{
\item{object}{eset}
}
\value{
default value of feature_plots
}
\description{
default feature_plots
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   autonomics.data::stemcomp.proteinratios \%>\%
   autonomics.plot::default_feature_plots()

   autonomics.data::stemcomp.soma \%>\%
   autonomics.plot::default_feature_plots()
}

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   autonomics.data::stemdiff.proteinratios \%>\%
   autonomics.plot::default_feature_plots()
}
if (require(billing.differentiation.data)){
   billing.differentiation.data::rna.voomcounts \%>\%
      default_feature_plots()
}

}
