% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_feature_plots}
\alias{default_feature_plots}
\title{default feature_plots}
\usage{
default_feature_plots(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
default value of feature_plots
}
\description{
default feature_plots
}
\examples{
if (require(autonomics.data)){
   require(magrittr)

   # STEM CELL COMPARISON
      autonomics.data::stemcomp.proteinratios \%>\%
      autonomics.plot::default_feature_plots()

      autonomics.data::stemcomp.soma \%>\%
      autonomics.plot::default_feature_plots()

   # GLUTAMINASE
      autonomics.data::glutaminase \%>\%
      autonomics.plot::default_feature_plots()
}
}
