% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_fvars}
\alias{default_fvars}
\title{default fvars}
\usage{
default_fvars(object)
}
\arguments{
\item{object}{eset}
}
\value{
character vector
}
\description{
default fvars
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   autonomics.data::billing2016 \%>\% autonomics.plot::default_fvars()
}
if (require(atkin.2014)){
   atkin.2014::soma \%>\% autonomics.plot::default_fvars()
}
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.counts
}
}
