% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_group_var}
\alias{default_group_var}
\title{Default group var}
\usage{
default_group_var(object)
}
\arguments{
\item{object}{eset}
}
\value{
default value of group_var
}
\description{
Default group var
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   autonomics.data::billing2016 \%>\% default_group_var()
}
if (require(atkin.2014)){
  atkin.2014::soma \%>\% autonomics.plot::default_group_var()
}
if (require(billing.differentiation.data)){
  billing.differentiation.data::rna.voomcounts \%>\%
     default_group_var()
}
}
