% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_line}
\alias{default_line}
\title{Default line value}
\usage{
default_line(object)
}
\arguments{
\item{object}{eset}
}
\value{
default line value
}
\description{
Default line value
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   autonomics.data::stemcomp.proteinratios \%>\% autonomics.plot::default_line()
}

# ATKIN.2014
if (require(atkin.2014)){
  atkin.2014::soma \%>\% autonomics.plot::default_line()
}

# STEM CELL DIFFERENTIATION
if (require(billing.differentiation.data)){
  billing.differentiation.data::rna.voomcounts \%>\%
     default_line()
}
}
