% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_shape_var}
\alias{default_shape_var}
\title{default color_var}
\usage{
default_shape_var(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
default value of color_var
}
\description{
default color_var
}
\examples{
if (require(autonomics.data)){
   require(magrittr)

   # STEM CELL COMPARISON
   autonomics.data::stemcomp.proteinratios \%>\% autonomics.plot::default_shape_var()

   # GLUTAMINASE
   autonomics.data::glutaminase \%>\% autonomics.plot::default_shape_var()

}
}
