% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_x}
\alias{default_x}
\title{Default x}
\usage{
default_x(object, feature_plots = default_feature_plots(object))
}
\arguments{
\item{object}{SummarizedExperiment}

\item{feature_plots}{feature plots}
}
\value{
default value of x
}
\description{
Default x
}
\examples{
if (require(autonomics.data)){
   require(magrittr)

   # STEM CELL COMPARISON
   autonomics.data::stemcomp.proteinratios \%>\% default_x()

   # GLUTAMINASE
   autonomics.data::glutaminase            \%>\% default_x()
}
}
