% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_zero_hline}
\alias{default_zero_hline}
\title{Default zero hline}
\usage{
default_zero_hline(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
logical
}
\description{
Default zero hline
}
\examples{
if (require(autonomics.data)){
   require(magrittr)

   # STEM CELL COMPARISON
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% default_zero_hline()

   # GLUTAMINASE
   object <- autonomics.data::glutaminase
   object \%>\% default_zero_hline()
}
}
