% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_zero_hline}
\alias{default_zero_hline}
\title{Default zero hline}
\usage{
default_zero_hline(object)
}
\arguments{
\item{object}{eset}
}
\value{
logical
}
\description{
Default zero hline
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   autonomics.data::stemcomp.proteinratios \%>\%
   autonomics.plot::default_zero_hline()
}

# STEM CELL DIFFERENTIATION
if (require(billing.differentiation.data)){
   billing.differentiation.data::rna.voomcounts \%>\%
      autonomics.plot::default_zero_hline()
}
}
