% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{gg_aesvar}
\alias{gg_aesvar}
\title{Get name of variable mapped to a aesthetic}
\usage{
gg_aesvar(b, aesthetic)
}
\arguments{
\item{b}{built ggplot (result of ggplot_build)}

\item{aesthetic}{ggplot aesthetic (string)}
}
\value{
character vector
}
\description{
Get name of variable mapped to a aesthetic
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios \%>\% magrittr::extract(1:9, )
   b <- object \%>\% autonomics.plot::plot_features('point') \%>\%
                   ggplot2::ggplot_build()
   b \%>\% autonomics.plot::gg_aesvar(aesthetic = 'color')
   b \%>\% autonomics.plot::gg_aesvar(aesthetic = 'x')
   b \%>\% autonomics.plot::gg_aesvar(aesthetic = 'shape')
   b <- object \%>\% autonomics.plot::plot_features('violin') \%>\%
                   ggplot2::ggplot_build()
   b \%>\% autonomics.plot::gg_aesvar(aesthetic = 'color')
   b \%>\% autonomics.plot::gg_aesvar(aesthetic = 'x')
   b \%>\% autonomics.plot::gg_aesvar(aesthetic = 'shape')
}
if (require(atkin.2014)){
   object <- atkin.2014::soma \%>\% magrittr::extract(1:10, )
   b <- object \%>\% autonomics.plot::plot_features('point',
                        x = 'time', color_var = 'condition', facet_var = 'subject_id',
                        line = TRUE, fvars = 'TargetFullName') \%>\%
                        ggplot2::ggplot_build()
   b \%>\% gg_aesvar(aesthetic = 'colour')
}
if (require(subramanian.2016)){
   b <- subramanian.2016::metabolon \%>\%
        magrittr::extract(1:10, ) \%>\%
        plot_features('boxplot',
                       color_var = 'condition',
                       group_var = 'condition',
                       line      = TRUE) \%>\%
        ggplot2::ggplot_build()
   aesthetic <- 'colour'
   gg_aesvar(b, aesthetic)
}
}
