% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{gg_rowvars}
\alias{gg_rowvars}
\alias{gg_colvars}
\alias{gg_nchar_rowvars}
\alias{gg_panelvars}
\title{Get column variables}
\usage{
gg_rowvars(b)

gg_colvars(b)

gg_nchar_rowvars(b)

gg_panelvars(b)
}
\arguments{
\item{b}{built ggplot (result of ggplot_build)}
}
\value{
vector with column variables
}
\description{
Get column variables from faceted ggplot object
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios \%>\% magrittr::extract(1:9, )
   b <- object \%>\% autonomics.plot::plot_features('point') \%>\% ggplot2::ggplot_build()
   b \%>\% autonomics.plot::gg_rowvars()
   b \%>\% autonomics.plot::gg_colvars()
   b \%>\% autonomics.plot::gg_nchar_rowvars()
   b \%>\% autonomics.plot::gg_panelvars()
   b \%>\% autonomics.plot::gg_panelvar_width()
}
}
