% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{lda}
\alias{lda}
\title{Linear discriminant analysis (LDA)}
\usage{
lda(object, na.impute = FALSE, ndim = 2, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{na.impute}{currently ignored. May be used in future.}

\item{ndim}{Number of linear discriminants to include in results}

\item{...}{only included to keep \code{\link[autonomics.plot]{project}} generic}
}
\value{
\preformatted{list(samples  =   sample scores,       # matrix:  nsample x ndim
     features =  feature loadings,     # matrix: nfeature x ndim
     var      = variance percentages   # vector:        1 x ndim
}
}
\description{
Returns \code{LDA} sample scores, feature loadings, and discriminant variances.
}
\details{
\code{LDA} relies on the (\code{NA}-unfriendly) \code{SVD}, so only \code{NA}-free features are included in the analysis.
}
\examples{
if (require(autonomics.data)){
   require(magrittr)

   # STEM CELL COMPARISON
      object <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
                 system.file(package = 'autonomics.data')     \%>\%
                 autonomics.import::read_proteingroups()
      object \%>\% autonomics.plot::lda() \%>\% str()
      \dontrun{ # Fails, as max dim length(subgroup) - 1
         object \%>\% autonomics.plot::lda(ndim = 3) \%>\% str()
       }

   # STEM CELL DIFFERENTIATION
      require(magrittr)
      object <- 'extdata/stemdiff/maxquant/proteinGroups.txt' \%>\%
                 system.file(package='autonomics.data')       \%>\%
                 autonomics.import::read_proteingroups()
      object \%>\% lda() \%>\% str()

   # GLUTAMINASE
      object <- 'extdata/glutaminase/glutaminase.xlsx'    \%>\%
                 system.file(package = 'autonomics.data') \%>\%
                 autonomics.import::read_metabolon()
      object \%>\% lda() \%>\% str()
}
}
\seealso{
\code{\link[autonomics.plot]{pca}},
\code{\link[autonomics.plot]{sma}},
\code{\link[autonomics.plot]{pls}}
}
\author{
Aditya Bhagwat, Laure Cougnaud
}
