% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{make_colors}
\alias{make_colors}
\title{Make fitting colors}
\usage{
make_colors(x, sep = autonomics.import::infer_design_sep(x, verbose = FALSE),
  show = FALSE)
}
\arguments{
\item{x}{colorvar levels vector}

\item{show}{logical}
}
\value{
color vectors (values = colors, names = colorvar levels)
}
\description{
Make fitting colors
}
