% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{make_composite_colors}
\alias{make_composite_colors}
\title{Make composite colors}
\usage{
make_composite_colors(object,
  color_var = autonomics.plot::default_color_var(object))
}
\arguments{
\item{object}{SummarizedExperiment}

\item{color_var}{svar mapped to color}
}
\value{
named character vector (elements = colors, names = color_var levels)
}
\description{
Make composite colors
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% autonomics.plot::make_composite_colors()
}
if (require(halama.2016)){
   object <- halama.2016::cell.metabolites
   object \%>\% autonomics.plot::make_composite_colors()
}
}
