% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{make_composite_colors}
\alias{make_composite_colors}
\title{Make composite colors}
\usage{
make_composite_colors(svalues,
  sep = autonomics.import::guess_sep(svalues), show = FALSE)
}
\arguments{
\item{svalues}{string vector}

\item{sep}{string}

\item{show}{TRUE/FALSE: show colors in pie plot?}
}
\value{
named string vector (elements = colors, names = color_var levels)
}
\description{
Make composite colors
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   svalues <- autonomics.data::glutaminase \%>\%
              autonomics.import::subgroup_levels()
   svalues \%>\% make_composite_colors(show = TRUE)
}
}
