% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{make_volcano_dt}
\alias{make_volcano_dt}
\title{Create volcano datatable}
\usage{
make_volcano_dt(object, ntop = 3)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{ntop}{number: how many top features (either FC wise or p wise) to be annotated}
}
\value{
data.table
}
\description{
Create volcano datatable
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% make_volcano_dt() \%>\% print()
}
}
