% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_heatmap.R
\name{plot_feature_heatmap}
\alias{plot_feature_heatmap}
\title{Plot feature heatmap}
\usage{
plot_feature_heatmap(object, filename = NULL, lowColor = "yellow",
  highColor = "blue1", midColor = "gray", width = 2, height = 4,
  fontsize = 0, ...)
}
\arguments{
\item{object}{SummarizedExperiment or EList}

\item{filename}{character vector}

\item{lowColor}{color for downregulation (Default color yellow)}

\item{highColor}{color for upregulation (Default color royal blue)}

\item{midColor}{color for midpoint (Default color gray)}

\item{width}{width  in inches (Default width 2 inches)}

\item{height}{height in inches (Default height 4 inches)}

\item{fontsize}{font size for feature name (Default size 0 for not showing feature names)}

\item{...}{further parameters handed through to \code{\link[ggplot2]{theme}}}
}
\description{
Plot feature heatmap
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   rnaseq_dir <- system.file('extdata/rnaseq', package = 'subramanian.2016')
   object <- autonomics.import::load_rnaseq_counts(rnaseq_dir) \%>\%
             autonomics.import::logcpm() \%>\%
             magrittr::extract(1:10, )
   object \%>\% autonomics.plot::plot_feature_heatmap()
   filename <- paste0(tempfile(), '.pdf') \%T>\% message()
   object \%>\% autonomics.plot::plot_feature_heatmap(filename = filename)
}
}
