% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_heatmap.R
\name{plot_feature_heatmap_with_pheatmap}
\alias{plot_feature_heatmap_with_pheatmap}
\title{Plot feature heatmap}
\usage{
plot_feature_heatmap_with_pheatmap(object, fvar = "feature_id",
  filename = NULL, width = 7, height = 7, cellheight = 8,
  fontsize = 6, treeheight_row = 0, treeheight_col = 0,
  cluster_cols = FALSE, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{fvar}{string}

\item{filename}{character vector}

\item{width}{width  in inches}

\item{height}{height in inches}

\item{cellheight}{heatmap cell height}

\item{fontsize}{fontsize}

\item{treeheight_row}{height of tree for rows}

\item{treeheight_col}{height of tree for col}

\item{cluster_cols}{clustering on cols}

\item{...}{passed to pheatmap::pheatmap}
}
\description{
Plot feature heatmap
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::glutaminase[1:50, ]
   object \%>\% autonomics.plot::plot_feature_heatmap_with_pheatmap(fvar='BIOCHEMICAL')
   filename <- tempfile() \%T>\% message()
   object \%>\% autonomics.plot::plot_feature_heatmap_with_pheatmap(filename = filename)
}
}
