% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_heatmap.R
\name{plot_feature_heatmap_with_pheatmap}
\alias{plot_feature_heatmap_with_pheatmap}
\title{Plot feature heatmap}
\usage{
plot_feature_heatmap_with_pheatmap(
  object,
  fvar = "feature_id",
  filename = NULL,
  width = 7,
  height = 7,
  cellheight = 8,
  fontsize = 6,
  treeheight_row = 0,
  treeheight_col = 0,
  cluster_cols = FALSE,
  ...
)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{fvar}{string}

\item{filename}{string}

\item{width}{number: width  in inches}

\item{height}{number: height in inches}

\item{cellheight}{number: heatmap cell height}

\item{fontsize}{number: fontsize}

\item{treeheight_row}{number: height of tree for rows}

\item{treeheight_col}{number: height of tree for col}

\item{cluster_cols}{TRUE or FALSE}

\item{...}{passed to pheatmap::pheatmap}
}
\description{
Plot feature heatmap
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::glutaminase[1:50, ]
   object \%>\% plot_feature_heatmap_with_pheatmap(fvar='BIOCHEMICAL')
   filename <- tempfile() \%T>\% message()
   object \%>\% plot_feature_heatmap_with_pheatmap(filename = filename)
}
}
