% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{plot_features}
\alias{plot_features}
\alias{plot_feature_bars}
\alias{plot_feature_profiles}
\alias{plot_feature_distributions}
\alias{plot_feature_boxes}
\title{Plot features}
\usage{
plot_features(object,
  geom = autonomics.plot::default_feature_plots(object)[1],
  x = autonomics.plot::default_x(object, geom[1]),
  color_var = autonomics.plot::default_color_var(object),
  color_values = autonomics.plot::default_color_values(object, color_var),
  shape_var = autonomics.plot::default_shape_var(object),
  group_var = autonomics.plot::default_group_var(object),
  txt_var = autonomics.plot::default_txt_var(object), facet_var = NULL,
  alpha_var = NULL, fvars = autonomics.plot::default_fvars(object),
  title = "", scales = "free_y", x_text_angle = 90,
  line = autonomics.plot::default_line(object),
  zero_hline = autonomics.plot::default_zero_hline(object), xlab = NULL,
  ylab = NULL, dodge_width = 0, verbose = FALSE, file = NULL,
  width = NULL, height = NULL, theme = ggplot2::theme_bw(),
  legend.position = "right")

plot_feature_bars(...)

plot_feature_profiles(...)

plot_feature_distributions(...)

plot_feature_boxes(...)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{geom}{'point', 'bar', 'violin', 'boxplot'}

\item{x}{svar mapped to x}

\item{color_var}{svar mapped to color}

\item{color_values}{named color vector (names = color_var levels, values = colors)}

\item{shape_var}{svar mapped to shape (only relevant for point)}

\item{group_var}{svar mapped to group}

\item{txt_var}{svar mapped to txt}

\item{facet_var}{svar on which to facet plot}

\item{alpha_var}{svar (logical) mapped to absence/presence of transparancy}

\item{fvars}{fvar used for plot annotation}

\item{title}{plot title}

\item{scales}{'free'}

\item{x_text_angle}{angle of text on x axis}

\item{line}{logical: should line be added?}

\item{zero_hline}{logical: should y=0 line be added?}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{dodge_width}{numeric}

\item{verbose}{logical}

\item{file}{string: file path}

\item{width}{numeric}

\item{height}{numeric}

\item{theme}{ggplot2::theme}

\item{legend.position}{position of legend}

\item{...}{only for backward compatibility to deprecated functions}
}
\value{
file path
}
\description{
Plot features
}
\examples{
require(magrittr)
result_dir <- tempdir() \%T>\% message()

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios \%>\% extract(1:10, )
   object \%>\% autonomics.plot::plot_features(geom = 'violin')
   object \%>\% autonomics.plot::plot_features(geom = 'point')
   object \%>\% autonomics.plot::plot_features(geom = 'boxplot')
   object \%>\% autonomics.plot::plot_features(
                 geom = 'boxplot',
                 file = paste0(result_dir, '/stemcomp_boxes.pdf'))

}

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios \%>\% extract(1:10, )
   object \%>\% autonomics.plot::plot_features('boxplot')
}

if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts \%>\% extract(1:10, )
   object \%>\% autonomics.plot::plot_features('point',   fvars = 'gene_name')
   object \%>\% autonomics.plot::plot_features('violin',  fvars = 'gene_name')
   object \%>\% autonomics.plot::plot_features('boxplot', fvars = 'gene_name')
}

# GLUTAMINASE
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase[1:9, ]
   color_values <- c(Control = 'blue', Vehicle = 'green', `Concentration 1` = 'orange',
                    `Concentration 2` = 'red')
   object$alpha <- object$TREATMENT \%in\% c(0,1)
   object \%>\% autonomics.plot::plot_features(
                'boxplot',
                 x = 'TIME_POINT', color_var = 'GROUP_DESCRIPTION', color_values = color_values,
                 alpha_var = 'alpha')
}

# SUBRAMANIAN.2016
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon[1:10, ]
   object \%>\% plot_features('boxplot', color_var = 'condition',
                             group_var = 'condition', line = TRUE)

   object <- subramanian.2016::rnaseq[1:10, ]
   object \%>\% plot_features('boxplot', color_var = 'condition', group_var = 'condition',
                             line = TRUE)

}

# ATKIN.2014
if (require(atkin.2014)){
   object <- atkin.2014::soma \%>\% magrittr::extract(1:10, )
   object \%>\% plot_features('point', x = 'time', color_var = 'condition',
                             facet_var = 'subject_id', group_var = 'subject_id', line = TRUE)
   object \%>\% plot_features('point', x = 'time', color_var = 'condition', facet_var = 'subject_id',
                             group_var = 'subject_id', line = TRUE, fvar = 'TargetFullName')
   object \%>\% plot_features('point', x = 'time', color_var = 'condition', facet_var = 'subject_id',
                             group_var = 'subject_id', line = TRUE, fvar = 'TargetFullName')
}

}
