% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlayed_sample_distributions.R
\name{plot_overlayed_sample_distributions}
\alias{plot_overlayed_sample_distributions}
\title{Plot overlayed sample distributions}
\usage{
plot_overlayed_sample_distributions(object,
  color_var = default_color_var(object),
  color_values = default_color_values(object, color_var))
}
\arguments{
\item{object}{SummarizedExperiment}

\item{color_var}{string: svar mapped to color}

\item{color_values}{string vector: values = colors, names = color_var levels}
}
\value{
ggplot2 object
}
\description{
Plot overlayed sample distributions
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% plot_overlayed_sample_distributions()
}
}
