% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions.R
\name{plot_sample_densities}
\alias{plot_sample_densities}
\alias{plot_sample_densities.SummarizedExperiment}
\title{Plot sample densities}
\usage{
plot_sample_densities(object, ...)

\method{plot_sample_densities}{SummarizedExperiment}(object,
  title = "Sample densities", ...)
}
\arguments{
\item{object}{matrix, SummarizedExperiment, or list of SummarizedExperiments}

\item{...}{ggplot2 aesthetics or parameters}

\item{title}{string}

\item{sdata}{dataframe with sample data}

\item{x}{string: svar mapped to x}

\item{group}{string: svar mapped to group}

\item{color}{string: svar mapped to color}

\item{facet}{string: svar mapped to facets}

\item{xlab}{string}

\item{ylab}{string}
}
\value{
ggplot2 object
}
\description{
Plot sample densities
}
\examples{
if (require(autonomics.data)){

   # STEM CELL COMPARISON
      # matrix
        require(magrittr)
        object <- stemcomp.proteinratios \%>\% autonomics.import::exprs()
        sdata  <- stemcomp.proteinratios \%>\% autonomics.import::sdata()
        plot_sample_densities(object, sdata)
        plot_sample_densities(object, sdata, color = subgroup, facet = subgroup)

     # SummarizedExperiment
        object <- stemcomp.proteinratios
        plot_sample_densities(object)
        plot_sample_densities(object, color = subgroup)

   # GLUTAMINASE
      # matrix
        object <- autonomics.import::exprs(glutaminase)
        sdata  <- autonomics.import::sdata(glutaminase)
        plot_sample_densities(object, sdata, color = subgroup, facet = CONCENTRATION)

      # SummarizedExperiment
        object <- glutaminase
        plot_sample_densities(object)
        plot_sample_densities(object, color = subgroup)

}
}
