% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions.R
\name{plot_sample_distributions}
\alias{plot_sample_distributions}
\title{Plot distribution across features for each sample}
\usage{
plot_sample_distributions(object, ..., x = NULL,
  color_var = default_color_var(object),
  color_values = default_color_values(object, color_var),
  displayed_features = NULL, jitter_features = FALSE,
  horizontal = TRUE, xlab = "", ylab = "",
  title = sprintf("Per sample distribution of all \%d features",
  nrow(object)))
}
\arguments{
\item{object}{SummarizedExperiment}

\item{...}{additonal SummarizedExperiments or matrices AND further parameters for \code{\link{core_ggplot_from_data}}}

\item{x}{string: sample variable mapped to x axis}

\item{color_var}{string: sample variable mapped to color}

\item{color_values}{string vector: color values vector (names = subgroups, values = colors)}

\item{displayed_features}{number/string vector: features to be displayed in the sample distributions (vector of numeric indexes or character \code{feature_id}s)}

\item{jitter_features}{TRUE/FALSE: whether to use geom_jitter or not for the displayed features}

\item{horizontal}{TRUE/FALSE: whether to lay the plot out horizontally (measurements on the x, sample names on the y axis)}

\item{xlab}{string: label of y axis}

\item{ylab}{string: label of y axis}

\item{title}{string: title}
}
\value{
ggplot2 object
}
\description{
The distributions are plotted as a combination of a violin plot and a box plot.
A separate violin is created for each value of 'x' (rows) and 'facet_var (column).
The box is colored per value of color_var
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios

   # Simple plot
     object \%>\% plot_sample_distrsibutions()

   # Same thing vertical
     object \%>\% plot_sample_distributions(horizontal = FALSE)

   # More complex using facetting
     object \%>\%
     plot_sample_distributions(
        2^autonomics.import::exprs(autonomics.data::stemcomp.proteinratios),
        facet2_var = c('Logarithmized', 'Raw'))

   # Yet more crazy explicitly marking the feature with the maximum value
     object \%>\%
     plot_sample_distributions(2^autonomics.import::exprs(object),
                               facet2_var         = c('Logarithmized', 'Raw'),
                               displayed_features = autonomics.import::exprs(object)        \%>\%
                                                    equals(autonomics.import::exprs(object) \%>\%
                                                    max(na.rm = TRUE))                      \%>\%
                                                    which(arr.ind = TRUE)                   \%>\%
                                                    magrittr::extract(,'row'))
}
}
\seealso{
\code{\link{core_ggplot_from_data}}
}
\author{
Aditya Bhagwat, Johannes Graumann
}
