% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions.R
\name{plot_sample_distributions}
\alias{plot_sample_distributions}
\title{Plot distribution across features for each sample}
\usage{
plot_sample_distributions(object, x = NULL,
  color_var = autonomics.plot::default_color_var(object),
  color_values = autonomics.plot::default_color_values(object,
  color_var), facet_var = NULL, displayed_features = NULL, ylab = "",
  xlab = "",
  title = sprintf("Per sample distribution of all \%d features",
  nrow(object)), file = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{x}{sample variable mapped to x axis}

\item{color_var}{sample variable mapped to color}

\item{color_values}{comolor values vector (names = subgroups, values = colors)}

\item{facet_var}{sample variable for faceting}

\item{displayed_features}{features to be displayed in the sample distributions (vector of numeric indexes or character \code{feature_id}s)}

\item{ylab}{label of y axis (character)}

\item{xlab}{label of y axis (character)}

\item{title}{title (character)}

\item{file}{file to which to print plot}
}
\value{
ggplot2 object
}
\description{
The distributions are plotted as a combination of a violin plot and a box plot.
A separate violin is created for each value of 'x' (rows) and 'facet_var (column).
The box is colored per value of color_var
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   ALL <- autonomics.data::ALL[, 1:30]
   plot_sample_distributions(object = ALL, x = 'cod', color_var = 'BT')

   # faceting works, but is not useful here, as there is no block variable
   plot_sample_distributions(object = ALL, x = 'cod',
                             facet_var = 'sex', color_var = 'BT')
}

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   autonomics.data::stemdiff.proteinratios \%>\%
   autonomics.plot::plot_sample_distributions()
}
if (require(billing.differentiation.data)){
   billing.differentiation.data::rna.voomcounts \%>\%
      plot_sample_distributions()
}
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% plot_sample_distributions()
}

# GLUTMINASE
if (require(autonomics.data)){
   autonomics.data::glutaminase \%>\%
   autonomics.plot::plot_sample_distributions(
      color_var    = 'GROUP_DESCRIPTION',
      color_values = c(Control = 'gray94', Vehicle = 'darkgrey',
                      `Concentration 1` = 'green', `Concentration 2` = 'forestgreen')
   )
}
}
\author{
Aditya Bhagwat
}
