% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_distributions.R
\name{plot_sample_violins}
\alias{plot_sample_violins}
\title{Plot sample violins}
\usage{
plot_sample_violins(object, ...)
}
\arguments{
\item{object}{SummarizedExperiment or matrix}

\item{x}{string: sample variable mapped to x axis}

\item{color_var}{string: sample variable mapped to color}

\item{color_values}{string vector: names = subgroups, values = colors}

\item{facet_var}{string: sample variable for faceting}

\item{displayed_features}{features to be displayed in the sample distributions (vector of numeric indexes or character \code{feature_id}s)}

\item{xlab}{label of y axis (character)}

\item{ylab}{label of y axis (character)}

\item{title}{title (character)}

\item{file}{file to which to print plot}
}
\value{
ggplot2 object
}
\description{
Plot sample violins
}
\examples{
if (require(autonomics.data)){

   # ALL
      require(magrittr)
      object <- autonomics.data::ALL[, 1:30]
      object \%>\% plot_sample_densities(x = 'cod', color_var = 'BT')

      # faceting works, but is not useful here, as there is no block variable
      object \%>\% plot_sample_distributions_v1(x = 'cod', facet_var = 'sex', color_var = 'BT')

   # STEM CELL COMPARISON
      object <- autonomics.data::stemcomp.proteinratios
      object \%>\% plot_sample_distributions_v1()

   # GLUTMINASE
      x <- autonomics.data::glutaminase
      x \%<>\% autonomics.import::exprs()
      object \%>\% plot_sample_distributions_v1()
}
}
\author{
Aditya Bhagwat
}
