% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Plot Venn/Euler diagram}
\usage{
plot_venn(
  x,
  color_values = if (assertive.properties::has_names(x)) {     make_colors(names(x)) }
    else {     make_gg_colors(1:length(x)) },
  euler = FALSE,
  filename = NULL,
  width = 7,
  height = width,
  title = NULL,
  scaled = FALSE,
  cat.dist = 0.3,
  margin = cat.dist,
  main.cex = 1.5,
  main.pos = c(0.5, 0.97),
  ...
)
}
\arguments{
\item{x}{named list of character vectors}

\item{color_values}{string vector: values = colors, names = x levels}

\item{euler}{TRUE/FALSE: whether to use eulerr::euler() rather than VennDiagram::venn.diagram()}

\item{filename}{string}

\item{width}{number: width  in inches}

\item{height}{number: height in inches}

\item{title}{string: figure title}

\item{scaled}{TRUE/FALSE: whether or not to scale}

\item{cat.dist}{number: category location (as distance from circle edge)}

\item{margin}{number}

\item{main.cex}{number: scaling factor for title size}

\item{main.pos}{number: title location}

\item{...}{passed to either VennDiagram::venn.diagram(...) or eulerr::plot.euler(...)}
}
\description{
Plot Venn/Euler diagram
}
\examples{
x <- list(A = c('apple', 'pear'), B = c('pear', 'orange'))
filename <- NULL
plot_venn(x)
plot_venn(x, euler = TRUE)
filename <- tempfile()
plot_venn(x, filename = filename)
}
