% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venns.R
\name{plot_venns}
\alias{plot_venns}
\title{Plot Venn diagrams}
\usage{
plot_venns(x, filename = NULL, width = 7, height = width,
  title = NULL, colors = autonomics.plot::make_gg_colors(1:length(x)))
}
\arguments{
\item{x}{named list of character vectors}

\item{filename}{character vector}

\item{width}{width  in inches}

\item{height}{height in inches}

\item{title}{figure title}
}
\description{
Plot Venn diagrams
}
\examples{
x <- list(A = c('apple', 'pear'), B = c('pear', 'orange'))
filename <- NULL
autonomics.plot::plot_venns(x, filename = NULL)
filename <- tempfile()
autonomics.plot::plot_venns(x, filename = filename)
}
