% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot volcano}
\usage{
plot_volcano(object, ntop = 3, nrow = NULL, legend_position = NULL,
  labeled_features = TRUE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{ntop}{n top features to be annotated}

\item{nrow}{number of rows in faceted plot}

\item{legend_position}{handed to \code{\link[ggplot2]{theme}} as 'legend.position'}

\item{labeled_features}{whether to label prominent features or not}
}
\value{
ggplot object
}
\description{
Plot volcano
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   autonomics.import::fdata(object) \%<>\% magrittr::extract(, 1:47)
   object \%>\% autonomics.plot::plot_volcano()

   object <- subramanian.2016::exiqon
   autonomics.import::fdata(object) \%<>\% magrittr::extract(, 1:38)
   object \%>\% autonomics.plot::plot_volcano()

   \dontrun{
      object = subramanian.2016::rnaseq
      autonomics.import::fdata(object) \%<>\% magrittr::extract(, 1:22)
      #file <- tempfile() \%>\% paste0('.png') \%T>\% message()
      #png(file, width = 3*480, height = 2*480)
      object \%>\% autonomics.plot::plot_volcano()
      #dev.off()
   }
}
}
